refactor(Accounts): move AccountUsageDetails to folder [YTFRONT-5361]
