+++ packages/ui/src/ui/pages/accounts/tabs/detailed-usage/AccountUsageDetails/AccountUsageDetails.scss
